% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_delay_days}
\alias{calculate_delay_days}
\title{Calculate Delay Days}
\usage{
calculate_delay_days(df)
}
\arguments{
\item{df}{A data frame containing work order data}
}
\value{
A data frame with a new column for delay days
}
\description{
This function calculates the number of days between created and completed dates.
}
